feat: Move final managers - metrics_manager and contextual_manager to core/managers/
